remove bogus -I/usr/include/slang from CPPFLAGS
Hardcoding -I/usr/include/slang in CPPFLAGS is bogus for
cross-compilation. With recent versions of slang, the headers are
installed in ${sysroot}/usr/include directly, so there is no need for
an additional flag. And if one was needed, it should be added by the
configure script, after detecting the right header location.

Signed-off-by: Alex Suykov <alex.suykov@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>