restore automatic menu/list-height in whiptail
Commit 3f5d1d27f09b ("don't modify popt args in whiptail") removed
replacement of the menu/list-height popt argument, which caused the
height to not be automatically calculated together with the window
height and width, when they are specified as zero.

Rework the code to provide the menu/list-height as an additional
integer argument to listBox and checkList instead of replacing the popt
argument.